refactor(domain/): 필요 없는 도메인 삭제
사용하지 않는 해시화된 비밀번호에 대한 도메인 삭제
