+++ src/main/java/woowacourse/shoppingcart/application/dto/ModifiedCustomerDto.java
@@ -7,7 +7,7 @@ import woowacourse.shoppingcart.domain.customer.Customer;
 import woowacourse.shoppingcart.domain.customer.Email;
 import woowacourse.shoppingcart.domain.customer.Gender;
 import woowacourse.shoppingcart.domain.customer.Name;
-import woowacourse.shoppingcart.domain.customer.NewPassword;
+import woowacourse.shoppingcart.domain.customer.Password;
 import woowacourse.shoppingcart.dto.AddressRequest;
 import woowacourse.shoppingcart.dto.ModifiedCustomerRequest;
 
@@ -44,7 +44,7 @@ public class ModifiedCustomerDto {
 
     public static Customer toModifiedCustomerDto(final ModifiedCustomerDto modifiedCustomerDto) {
         return new Customer(new Email(modifiedCustomerDto.getEmail()),
-                new NewPassword(modifiedCustomerDto.getPassword()),
+                new Password(modifiedCustomerDto.getPassword()),
                 modifiedCustomerDto.getProfileImageUrl(), new Name(modifiedCustomerDto.getName()),
                 Gender.form(modifiedCustomerDto.getGender()),
                 new Birthday(modifiedCustomerDto.getBirthday()), new Contact(modifiedCustomerDto.getContact()),
