--- src/main/java/woowacourse/shoppingcart/domain/customer/ExistedPassword.java
@@ -1,20 +1,0 @@
-package woowacourse.shoppingcart.domain.customer;
-
-public class ExistedPassword implements Password {
-
-    private final String value;
-
-    public ExistedPassword(final String value) {
-        this.value = value;
-    }
-
-    @Override
-    public boolean isSamePassword(String hashedPassword) {
-        return this.value.equals(hashedPassword);
-    }
-
-    @Override
-    public String getPassword() {
-        return value;
-    }
-}
