+++ src/test/java/woowacourse/shoppingcart/dao/CustomerFixture.java
@@ -10,7 +10,7 @@ import woowacourse.shoppingcart.domain.customer.Customer;
 import woowacourse.shoppingcart.domain.customer.Email;
 import woowacourse.shoppingcart.domain.customer.Gender;
 import woowacourse.shoppingcart.domain.customer.Name;
-import woowacourse.shoppingcart.domain.customer.NewPassword;
+import woowacourse.shoppingcart.domain.customer.Password;
 import woowacourse.shoppingcart.domain.customer.Terms;
 
 public class CustomerFixture {
@@ -21,13 +21,13 @@ public class CustomerFixture {
             new AddressDto("a", "b", "12345"), true);
 
     public static final Customer updatedTommyDto = new Customer(1L, new Email("her0807@naver.com"),
-            new NewPassword("password1!"),
+            new Password("password1!"),
             "example.com", new Name("토미"), Gender.MALE, new Birthday("1988-08-07"),
             new Contact("01987654321"),
             new FullAddress("d", "e", "54321"), new Terms(true));
 
     public static final Supplier<Customer> tommyCreator = () -> new Customer(1L, new Email("her0807@naver.com"),
-            new NewPassword("password1!"),
+            new Password("password1!"),
             "example.com", new Name("토미"), Gender.MALE, new Birthday("1988-08-07"),
             new Contact("12345678910"),
             new FullAddress("a", "b", "12345"), new Terms(true));
